Partial implementation of dcmtk test program

+++ src/plastimatch/segment_body.cxx
@@ -209,7 +209,7 @@ get_largest_connected_component (UCharImageType::Pointer i2)
 	if (cc_pct > ccs_percent_thresh) {
 	    label_upper_thresh = ccs + 1;
 	    printf ("CC %d has size %d (%f)\n", ccs, 
-		rel_filter->GetSizeOfObjectsInPixels()[ccs],
+		(int) rel_filter->GetSizeOfObjectsInPixels()[ccs],
 		cc_pct
 	    );
 	} else {
